// frontend/src/app/components/pages-sidebar/rilevazioni-sito/rilevazioni-sito-dto.ts
export class RilevazioniSitoDto {
    idmisurazione?: any;
    numcodsito?: any;
    protocollo?: any;
    sistema?: any;
    azimuth?: any;
    altcentrele?: any;
    antenna?: any;
    altezzaantenna?: any;
    guadagnoantenna?: any;
    tiltmecca?: any;
    titltelettr?: any;
    numerotrx?: any;
    potmaxsingtrx?: any;
    potmaxtotant?: any;
    attcei?: any;
    alpha24?: any;
    fattriduz?: any;
    pottotantfattrid?: any;
}